+++ matchbox/http/ipxe_test.go
@@ -17,7 +17,7 @@ func TestIPXEInspect(t *testing.T) {
 	h := ipxeInspect()
 	w := httptest.NewRecorder()
 	req, _ := http.NewRequest("GET", "/", nil)
-	h.ServeHTTP(context.Background(), w, req)
+	h.ServeHTTP(w, req)
 	assert.Equal(t, http.StatusOK, w.Code)
 	assert.Equal(t, ipxeBootstrap, w.Body.String())
 }
@@ -29,7 +29,7 @@ func TestIPXEHandler(t *testing.T) {
 	ctx := withProfile(context.Background(), fake.Profile)
 	w := httptest.NewRecorder()
 	req, _ := http.NewRequest("GET", "/", nil)
-	h.ServeHTTP(ctx, w, req)
+	h.ServeHTTP(w, req.WithContext(ctx))
 	// assert that:
 	// - the Profile's NetBoot config is rendered as an iPXE script
 	expectedScript := `#!ipxe
@@ -47,7 +47,7 @@ func TestIPXEHandler_MissingCtxProfile(t *testing.T) {
 	h := srv.ipxeHandler()
 	w := httptest.NewRecorder()
 	req, _ := http.NewRequest("GET", "/", nil)
-	h.ServeHTTP(context.Background(), w, req)
+	h.ServeHTTP(w, req)
 	assert.Equal(t, http.StatusNotFound, w.Code)
 }
 
@@ -59,7 +59,7 @@ func TestIPXEHandler_RenderTemplateError(t *testing.T) {
 	ctx := withProfile(context.Background(), &storagepb.Profile{Boot: nil})
 	w := httptest.NewRecorder()
 	req, _ := http.NewRequest("GET", "/", nil)
-	h.ServeHTTP(ctx, w, req)
+	h.ServeHTTP(w, req.WithContext(ctx))
 	assert.Equal(t, http.StatusNotFound, w.Code)
 }
 
@@ -70,7 +70,7 @@ func TestIPXEHandler_WriteError(t *testing.T) {
 	ctx := withProfile(context.Background(), fake.Profile)
 	w := NewUnwriteableResponseWriter()
 	req, _ := http.NewRequest("GET", "/", nil)
-	h.ServeHTTP(ctx, w, req)
+	h.ServeHTTP(w, req.WithContext(ctx))
 	assert.Equal(t, http.StatusInternalServerError, w.Code)
 	assert.Empty(t, w.Body.String())
 }